fix websocket proxy
