+++ apps/portal-web/package.json
@@ -91,7 +91,8 @@
     "postcss": "8.4.31",
     "ts-log": "2.2.5",
     "webpack": "5.89.0",
-    "fs-extra": "11.1.1"
+    "fs-extra": "11.1.1",
+    "@types/http-proxy": "1.17.14"
   },
   "peerDependencies": {
     "monaco-editor": "0.44.0"
